// client/src/api/model/playerEquipment.ts
/**
 * Generated by orval v8.4.0 🍺
 * Do not edit manually.
 * Genesys API Documentation
 * Interactive API documentation
 * OpenAPI spec version: 1.0
 */
import type { ItemTemplate } from './itemTemplate';

export interface PlayerEquipment {
  weapons: ItemTemplate[];
  equippedArmor: ItemTemplate;
  otherGear: ItemTemplate[];
}