// client/src/api/model/equipmentType.ts
/**
 * Generated by orval v8.0.0 🍺
 * Do not edit manually.
 * Genesys API Documentation
 * Interactive API documentation
 * OpenAPI spec version: 1.0
 */

export type EquipmentType = typeof EquipmentType[keyof typeof EquipmentType];


export const EquipmentType = {
  Weapon: 'Weapon',
  Armor: 'Armor',
  Gear: 'Gear',
} as const;